+++ source/Public/Add-DeploymentAudienceMember.ps1
@@ -5,11 +5,15 @@ function Add-DeploymentAudienceMember {
     .DESCRIPTION
         This function will check if the deployments audiences have the devices as members, and if not they will be added to the audience.
     .EXAMPLE
-        Add-DeploymentAudienceMember -azureDeviceIDs ("ID1","ID2") -updateAudienceID <AudienceID>
+        Add-DeploymentAudienceMember -azureDeviceIDs ("ID1","ID2") -audienceID <AudienceID>
     .PARAMETER azureDeviceIDs
         The Azure Device IDs to add to the audience.
-    .PARAMETER updateAudienceID
+    .PARAMETER audienceID
         The Update Audience ID to add the members to.
+    .PARAMETER policyID
+        The Update Policy ID to add the members to.
+    .NOTES
+        You can specify either the audienceID or policyID parameter, if both are specified the audienceID will be used.
     #>
     [CmdletBinding()]
     param (
@@ -17,11 +21,17 @@ function Add-DeploymentAudienceMember {
         [array]
         $azureDeviceIDs,
         # The Update Audience ID
-        [Parameter(Mandatory = $true)]
+        [Parameter(Mandatory = $false)]
         [string]
-        $updateAudienceID
+        $audienceID,
+        [Parameter(Mandatory = $false)]
+        [string]
+        $policyID
     )
     begin {
+        if ([String]::IsNullOrEmpty($audienceID) -and ([String]::IsNullOrEmpty($policyID))) {
+            throw "You must specify either the audienceID or policyID parameter."
+        }
         # Create the param body base
         $paramBody = @{
             addMembers = @(
@@ -29,7 +39,12 @@ function Add-DeploymentAudienceMember {
         }
     }
     process {
-        $updateAudienceMembers = Get-DeploymentAudienceMember -policyID $updateAudienceID
+        IF (-Not([String]::IsNullOrEmpty($audienceID))) {
+            $updateAudienceMembers = Get-DeploymentAudienceMember -audienceID $audienceID
+        }
+        elseif (([String]::IsNullOrEmpty($policyID))) {
+            $updateAudienceMembers = Get-DeploymentAudienceMember -policyID $policyID
+        }
         foreach ($id in $azureDeviceIDs) {
             IF (-Not($updateAudienceMembers.id -contains $id)) {
                 $memberObject = @{
@@ -44,7 +59,7 @@ function Add-DeploymentAudienceMember {
         IF ($paramBody.addMembers.Count -ge 1) {
             Invoke-MgGraphRequest `
                 -Method POST `
-                -Uri "https://graph.microsoft.com/beta/admin/windows/updates/deploymentAudiences('$updateAudienceID')/updateAudience" `
+                -Uri "https://graph.microsoft.com/beta/admin/windows/updates/deploymentAudiences('$audienceID')/updateAudience" `
                 -Body $paramBody
         }
     }